refactor(cli): extract test subcommand into commands/test.rs
Move run_test into commands/test.rs. Its output/diagnostic helpers
(build_stop_reason_details, write_junit_xml, write_config_error_outputs,
simple_regex_is_match) are shared with other commands, so they stay in
main.rs as pub(crate); build_system_bus + run_start likewise.

Pure move — build + clippy clean.
